rename extreactor modules so it makes more sense with new arch
