+++ ibdl-extractors/src/imageboards/e621/pool.rs
@@ -5,7 +5,7 @@ use ibdl_common::{
     serde_json, ImageBoards,
 };
 
-use crate::{error::ExtractorError, websites::PoolExtract};
+use crate::{error::ExtractorError, imageboards::PoolExtract};
 
 use super::{models::E621PoolList, E621Extractor};
