+++ ibdl-extractors/src/prelude.rs
@@ -1,8 +1,8 @@
-pub use crate::websites::AsyncFetch;
-pub use crate::websites::Auth;
-pub use crate::websites::Extractor;
-pub use crate::websites::ExtractorThreadHandle;
-pub use crate::websites::MultiWebsite;
-pub use crate::websites::PoolExtract;
-pub use crate::websites::PostFetchAsync;
-pub use crate::websites::SinglePostFetch;
+pub use crate::imageboards::AsyncFetch;
+pub use crate::imageboards::Auth;
+pub use crate::imageboards::Extractor;
+pub use crate::imageboards::ExtractorThreadHandle;
+pub use crate::imageboards::MultiWebsite;
+pub use crate::imageboards::PoolExtract;
+pub use crate::imageboards::PostFetchAsync;
+pub use crate::imageboards::SinglePostFetch;
